chore: internalからdomainへ改名
